Move some e2e X tests
