+++ tests/e2e/ui/popover/test_popover.py
